+++ CHANGES.md
@@ -1,2 +1,3 @@
 # 1.1.0 ()
 * Added `RGB.Sub()` and `RGBA.Sub()`.
+* Added `SRGB.Alpha()`, `SRGBA.SRGB()` `RGB.Alpha()` and `RGBA.RGB()`.